requirements.txt
aiogram==3.5.0
python-dotenv==1.0.1
openai==1.13.3
langdetect==1.0.9
aiosqlite==0.20.0
pydub==0.25.1